Update Simulator
We added a second LED, so one lights up when we need to put air in the ball and one lights up when we need to take air out. We also added motor controls, so the motor is on for the duration of the contraction and off for the break between contractions.